+++ src/main/kotlin/io/openfuture/api/controller/api/FrontendPropertiesApiController.kt
@@ -1,7 +1,8 @@
 package io.openfuture.api.controller.api
 
 import io.openfuture.api.component.web3.Web3Wrapper
-import io.openfuture.api.domain.FrontendPropertiesDto
+import io.openfuture.api.config.propety.EthereumProperties
+import io.openfuture.api.domain.property.FrontendPropertiesDto
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
@@ -11,13 +12,14 @@ import org.web3j.spring.autoconfigure.Web3jProperties
 @RequestMapping("/api/properties")
 class FrontendPropertiesApiController(
         private val web3: Web3Wrapper,
-        private val properties: Web3jProperties
+        private val properties: EthereumProperties,
+        private val web3Properties: Web3jProperties
 ) {
 
     @GetMapping
     fun get(): FrontendPropertiesDto {
         val version = web3.getNetVersion()
-        return FrontendPropertiesDto(properties.clientAddress, version)
+        return FrontendPropertiesDto(web3Properties.clientAddress, version, properties.openTokenAddress!!)
     }
 
 }