+++ src/main/kotlin/io/openfuture/api/domain/property/FrontendPropertiesDto.kt
@@ -1,0 +1,7 @@
+package io.openfuture.api.domain.property
+
+data class FrontendPropertiesDto(
+        val networkAddress: String,
+        val networkVersion: String,
+        val openTokenAddress: String
+)