+++ src/test/kotlin/io/openfuture/api/controller/api/OpenKeyApiControllerTests.kt
@@ -1,6 +1,7 @@
 package io.openfuture.api.controller.api
 
 import io.openfuture.api.config.ControllerTests
+import io.openfuture.api.domain.auth.OpenKeyDto
 import io.openfuture.api.domain.scaffold.GenerateOpenKeyRequest
 import io.openfuture.api.entity.auth.OpenKey
 import io.openfuture.api.entity.auth.Role
@@ -9,17 +10,16 @@ import org.mockito.BDDMockito.given
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.http.HttpHeaders.AUTHORIZATION
 import org.springframework.http.MediaType.APPLICATION_JSON_UTF8
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.*
 import java.util.*
 
 @WebMvcTest(OpenKeyApiController::class)
-class OpenKeyApiControllerTest : ControllerTests() {
+class OpenKeyApiControllerTests : ControllerTests() {
 
     @Test
     fun generateTokenTest() {
-        val openKey = createOpenKey(setOf(Role("ROLE_DEPLOY")))
+        val openKey = createOpenKey(setOf(Role("ROLE_MASTER")))
         val request = GenerateOpenKeyRequest(Date())
         val newOpenKey = OpenKey(openKey.user, request.expireDate)
         val requestJson = objectMapper.writeValueAsString(request)
@@ -43,7 +43,7 @@ class OpenKeyApiControllerTest : ControllerTests() {
     }
 
     @Test
-    fun generateTokenWhenOpenTokenIsNotFoundShouldRedirectToIndexPage() {
+    fun generateTokenWhenOpenTokenIsNotFoundShouldRedirectToIndexPageTest() {
         val invalidToken = "not_valid_token"
 
         given(keyService.find(invalidToken)).willReturn(null)
@@ -53,6 +53,35 @@ class OpenKeyApiControllerTest : ControllerTests() {
 
                 .andExpect(status().is3xxRedirection)
                 .andExpect(redirectedUrl("http://localhost/"))
+    }
+
+    @Test
+    fun getAllTest() {
+        val openKey = createOpenKey(setOf(Role("ROLE_MASTER")))
+
+        given(keyService.find(openKey.value)).willReturn(openKey)
+        given(keyService.getAll(openKey.user)).willReturn(listOf(openKey))
+
+        mvc.perform(get("/api/keys")
+                .header(AUTHORIZATION, openKey.value))
+
+                .andExpect(status().isOk)
+                .andExpect(content().json(objectMapper.writeValueAsString(listOf(OpenKeyDto(openKey))), true))
+    }
+
+    @Test
+    fun disableTest() {
+        val openKey = createOpenKey(setOf(Role("ROLE_MASTER")))
+        val expectedOpenKey = createOpenKey(setOf(Role("ROLE_MASTER"))).apply { enabled = false }
+
+        given(keyService.find(openKey.value)).willReturn(openKey)
+        given(keyService.disable(openKey.value)).willReturn(expectedOpenKey)
+
+        mvc.perform(delete("/api/keys/" + openKey.value)
+                .header(AUTHORIZATION, openKey.value))
+
+                .andExpect(status().isOk)
+                .andExpect(content().json(objectMapper.writeValueAsString(OpenKeyDto(expectedOpenKey)), true))
     }
 
 }